// src/types/api.generated.ts
export interface components {
  schemas: {
    Game: {
      readonly id: number;
      readonly board: number[][];
      winner: components["schemas"]["ReducedUser"];
      first_player: components["schemas"]["ReducedUser"];
      second_player: components["schemas"]["ReducedUser"];
      /** Format: date-time */
      readonly created: string;
      readonly status: components["schemas"]["StatusEnum"];
    };
    LoginInput: {
      username: string;
      password: string;
    };
    LoginOutput: {
      token: string;
      username: string;
      id: number;
    };
    MakeMove: {
      row: number;
      col: number;
    };
    PaginatedGameList: {
      /** @example 123 */
      count?: number;
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?offset=400&limit=100
       */
      next?: string | null;
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?offset=200&limit=100
       */
      previous?: string | null;
      results?: components["schemas"]["Game"][];
    };
    PaginatedUserList: {
      /** @example 123 */
      count?: number;
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?offset=400&limit=100
       */
      next?: string | null;
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?offset=200&limit=100
       */
      previous?: string | null;
      results?: components["schemas"]["User"][];
    };
    ReducedUser: {
      readonly id: number;
      /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
      readonly username: string;
    };
    RegisterInput: {
      username: string;
      password: string;
    };
    /** @enum {string} */
    StatusEnum: "open" | "progress" | "finished";
    User: {
      readonly id: number;
      /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
      readonly username: string;
      game_count: number;
      /** Format: double */
      win_rate: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}